+++ R/aaa.R
@@ -13,7 +13,7 @@ globalVariables(c(".","n",".resid","null.deviance","r_environment",".fitted","Pr
 #' @importFrom dplyr mutate_at mutate_all summarise_at summarise_all arrange arrange_at select select_at filter mutate mutate_ funs group_by group_by_ summarise summarise_ slice as_data_frame bind_cols bind_rows desc first last min_rank data_frame inner_join arrange_at group_by_at ungroup
 #' @importFrom rlang .data parse_exprs :=
 #' @importFrom magrittr %>% %<>% %T>% set_colnames set_rownames set_names extract2
-#' @importFrom tidyr spread_ gather_ gather
+#' @importFrom tidyr spread gather
 #' @importFrom lubridate now
 #' @importFrom pryr where
 #' @importFrom DiagrammeR DiagrammeROutput renderDiagrammeR DiagrammeR mermaid