+++ NEWS.md
@@ -1,4 +1,8 @@
-# CHANGES IN radiant.model 0.8.7
+# CHANGES IN radiant.model 0.8.7.2
+
+* Upgraded tidyr dependency to 0.7
+
+# CHANGES IN radiant.model 0.8.7.1
 
 * Upgraded dplyr dependency to 0.7.1